--- .clusterfuzzlite/project.yaml
@@ -1,1 +1,0 @@
-language: python
