+++ .clusterfuzzlite/seed_corpora/apps/common/search/query_parser/boolean_true
@@ -1,0 +1,1 @@
+archived:true
