+++ .clusterfuzzlite/seed_corpora/apps/common/search/query_parser/compound_query
@@ -1,0 +1,1 @@
+language:"C++" AND stars:>=10
