+++ .clusterfuzzlite/seed_corpora/apps/common/search/query_parser/date_value
@@ -1,0 +1,1 @@
+created:2024-01-01
