+++ .clusterfuzzlite/seed_corpora/apps/common/search/query_parser/field_value
@@ -1,0 +1,1 @@
+language:python
