+++ .clusterfuzzlite/seed_corpora/apps/common/search/query_parser/malformed_syntax
@@ -1,0 +1,1 @@
+::: malformed
