+++ .clusterfuzzlite/seed_corpora/apps/common/search/query_parser/numeric_range
@@ -1,0 +1,1 @@
+stars:>100
