+++ .clusterfuzzlite/seed_corpora/apps/common/search/query_parser/quoted_string
@@ -1,0 +1,1 @@
+query:"hello world"
