+++ .clusterfuzzlite/seed_corpora/apps/slack/common/text/markdown_link
@@ -1,0 +1,1 @@
+[OWASP Nest](https://nest.owasp.dev)
