+++ .clusterfuzzlite/seed_corpora/apps/slack/common/text/mixed_markdown
@@ -1,0 +1,1 @@
+*bold* and <script>alert(1)</script>
