+++ .clusterfuzzlite/seed_corpora/apps/slack/common/text/slack_link
@@ -1,0 +1,1 @@
+<https://owasp.org|OWASP>
