+++ backend/Makefile
@@ -5,7 +5,9 @@ include backend/apps/nest/Makefile
 include backend/apps/owasp/Makefile
 include backend/apps/slack/Makefile
 
-.PHONY: audit-backend-dependencies compile-backend-requirements fetch-nest-dump sbom-backend-image upload-nest-dump
+.PHONY: audit-backend-dependencies build-cluster-fuzz-lite build-cluster-fuzz-lite-fuzzers \
+	clean-cluster-fuzz-lite compile-backend-requirements fetch-nest-dump sbom-backend-image \
+	test-cluster-fuzz-lite upload-nest-dump
 
 ##@ Backend
 
@@ -254,6 +256,53 @@ test-fuzz: ## Run fuzz tests (REST API and GraphQL)
 	@COMPOSE_BAKE=true DOCKER_BUILDKIT=1 \
 	docker compose --project-name nest-fuzz -f docker-compose/fuzz/compose.yaml up --build --remove-orphans --abort-on-container-exit db cache backend graphql
 
+CFL_DOCKERFILE ?= .clusterfuzzlite/Dockerfile
+CFL_IMAGE ?= nest-cluster-fuzz-lite
+CFL_OUT_DIR ?= $(CURDIR)/.clusterfuzzlite/out
+CFL_PLATFORM ?= linux/amd64
+CFL_SANITIZER ?= address
+CFL_SECONDS ?= 60
+CFL_TARGETS ?= text query_parser
+
+build-cluster-fuzz-lite: ## Build the ClusterFuzzLite Docker image
+	@docker build \
+		--platform $(CFL_PLATFORM) \
+		-f $(CFL_DOCKERFILE) \
+		-t $(CFL_IMAGE) \
+		$(CURDIR)
+
+build-cluster-fuzz-lite-fuzzers: build-cluster-fuzz-lite ## Compile ClusterFuzzLite fuzz targets
+	@mkdir -p $(CFL_OUT_DIR)
+	@docker run \
+		--platform $(CFL_PLATFORM) \
+		--rm \
+		-e ARCH=x86_64 \
+		-e FUZZING_ENGINE=libfuzzer \
+		-e OUT=/out \
+		-e SANITIZER=$(CFL_SANITIZER) \
+		-v $(CFL_OUT_DIR):/out \
+		$(CFL_IMAGE) \
+		bash build.sh
+
+test-cluster-fuzz-lite: build-cluster-fuzz-lite-fuzzers ## Run ClusterFuzzLite fuzz targets locally
+	@for target in $(CFL_TARGETS); do \
+		echo "Running ClusterFuzzLite target: $$target ($(CFL_SECONDS)s)"; \
+		docker run \
+		--platform $(CFL_PLATFORM) \
+		--rm \
+		-e ARCH=x86_64 \
+		-e FUZZING_ENGINE=libfuzzer \
+		-e OUT=/out \
+		-e SANITIZER=$(CFL_SANITIZER) \
+		-v $(CFL_OUT_DIR):/out \
+		$(CFL_IMAGE) \
+		/out/$$target -max_total_time=$(CFL_SECONDS); \
+	done
+
+clean-cluster-fuzz-lite: ## Remove locally built ClusterFuzzLite artifacts
+	@rm -rf $(CFL_OUT_DIR)
+	@docker image rm -f $(CFL_IMAGE) >/dev/null 2>&1 || true
+
 compile-backend-requirements: ## Regenerate hashed pip requirements from backend/requirements/*.in
 	@docker run \
 		--rm \
@@ -265,6 +314,7 @@ compile-backend-requirements: ## Regenerate hashed pip requirements from backend
 		$$(grep -E '^FROM python:' docker/backend/Dockerfile.local | sed 's/^FROM //; s/ AS .*//' | head -1) \
 		sh -c 'python -m pip install --no-warn-script-location --quiet pip-tools && \
 		python -m piptools compile --no-strip-extras --generate-hashes --output-file=backend/requirements/build.txt backend/requirements/build.in && \
+		python -m piptools compile --no-strip-extras --generate-hashes --output-file=backend/requirements/cluster-fuzz-lite.txt backend/requirements/cluster-fuzz-lite.in && \
 		python -m piptools compile --no-strip-extras --generate-hashes --output-file=backend/requirements/pre-commit.txt backend/requirements/pre-commit.in'
 
 audit-backend-dependencies: ## Audit backend Python dependencies
