+++ backend/apps/slack/__init__.py
@@ -1,7 +1,1 @@
 """OWASP Nest Slack app."""
-
-from apps.slack import (
-    actions,
-    commands,
-    events,
-)
