+++ backend/apps/slack/utils/__init__.py
@@ -1,13 +1,11 @@
 """Slack app utils package."""
 
+from apps.slack.common.text import escape, format_links_for_slack, strip_markdown
 from apps.slack.utils.format import (
-    escape,
-    format_links_for_slack,
     get_gsoc_projects,
     get_news_data,
     get_posts_data,
     get_sponsors_data,
     get_staff_data,
     get_text,
-    strip_markdown,
 )
