+++ backend/apps/slack/utils/format.py
@@ -3,9 +3,7 @@
 from __future__ import annotations
 
 import logging
-import re
 from functools import lru_cache
-from html import escape as escape_html
 from typing import TYPE_CHECKING
 from urllib.parse import urljoin
 
@@ -18,38 +16,9 @@ from lxml import html
 from requests.exceptions import RequestException
 
 from apps.common.constants import NL, OWASP_NEWS_URL
+from apps.slack.common.text import strip_markdown
 
 logger: logging.Logger = logging.getLogger(__name__)
-
-
-def escape(content: str) -> str:
-    """Escape HTML content.
-
-    Args:
-        content (str): The HTML content to escape.
-
-    Returns:
-        str: The escaped HTML content.
-
-    """
-    return escape_html(content, quote=False)
-
-
-def format_links_for_slack(text: str) -> str:
-    """Convert Markdown links to Slack markdown link format.
-
-    Args:
-        text (str): The input text that may include Markdown links.
-
-    Returns:
-        str: Text with Markdown links converted to Slack markdown links.
-
-    """
-    if not text:
-        return text
-
-    markdown_link_pattern = re.compile(r"\[([^\]]+)\]\((https?://[^\s)]+)\)")
-    return markdown_link_pattern.sub(r"<\2|\1>", text)
 
 
 @lru_cache
@@ -230,17 +199,3 @@ def get_text(blocks: tuple) -> str:
                     text.append(block["text"]["text"])
 
     return NL.join(text).strip()
-
-
-def strip_markdown(text: str) -> str:
-    """Strip markdown formatting.
-
-    Args:
-        text (str): The text with markdown formatting.
-
-    Returns:
-        str: The text with markdown formatting removed.
-
-    """
-    slack_link_pattern = re.compile(r"<(https?://[^|]+)\|([^>]+)>")
-    return slack_link_pattern.sub(r"\2 (\1)", text).replace("*", "")
