+++ backend/requirements/cluster-fuzz-lite.in
@@ -1,0 +1,1 @@
+pyparsing==3.3.2
