--- backend/requirements/fuzz.in
@@ -1,5 +1,0 @@
-atheris==3.0.0
-lxml
-pyparsing
-pyyaml
-requests
